fix: modify document filename
