--- pages/_documents.tsx
@@ -1,13 +1,0 @@
-import { Head, Html, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head>{/* Unable to take effect, to be resolved */}</Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
